Refactor(extension) Browser language
Moved logic from global space to extension space (inside define(), this avoid situations where another function with the same name exists in the global space) and created a method to wrap that logic (created a unit, this also increases redability and allows for easier extraction in the future).

Created a angular framework rootscope variable called browserLanguage in the controller method. The variable will be passed in the paint method 1st argument

Opted to use split() method for performance (avoid having to create variables and 2 functions calls)